+++ src/interpreter.c
@@ -486,6 +486,12 @@ w_value_t w_value_index(w_ctx_t *ctx, w_value_t *left, w_value_t *right) {
 						CMD(string_dup_mut);
 					if(w_streqc(str, "dup"))
 						CMD(string_dup);
+					if(w_streqc(str, "split"))
+						CMD(string_split);
+					if(w_streqc(str, "reverse!"))
+						CMD(string_reverse_mut);
+					if(w_streqc(str, "reverse"))
+						CMD(string_reverse);
 					char *cstr = w_cstring(str);
 					w_status_err(ctx->status, w_error_new((w_filepos_t){}, "No member '%s' in string.", cstr));
 					free(cstr);
@@ -538,6 +544,14 @@ w_value_t w_value_index(w_ctx_t *ctx, w_value_t *left, w_value_t *right) {
 						CMD(list_fill_mut);
 					if(w_streqc(str, "fill"))
 						CMD(list_fill);
+					if(w_streqc(str, "dup!"))
+						CMD(list_dup_mut);
+					if(w_streqc(str, "dup"))
+						CMD(list_dup);
+					if(w_streqc(str, "reverse!"))
+						CMD(list_reverse_mut);
+					if(w_streqc(str, "reverse"))
+						CMD(list_reverse);
 					char *cstr = w_cstring(str);
 					w_status_err(ctx->status, w_error_new((w_filepos_t){}, "No member '%s' in list.", cstr));
 					free(cstr);